chore: setup splitting PR changed files table, moves original table into it's own sub folder.
