+++ src/pages/PullRequestPage/subroute/FilesChangedTab/FilesChanged/Table/index.js
@@ -1,0 +1,1 @@
+export { default } from './Table'
